requirements_qa.txt
Django==2.0.8
djangorestframework==3.8.2
musicbrainzngs==0.6
python-memcached==1.59
pytz==2018.5
six==1.11.0
